+++ lib/karo/cache.rb
@@ -13,10 +13,13 @@ module Karo
 	  def search(name="")
 	    configuration = Config.load_configuration(options)
 	    path = File.join(configuration["path"], "shared/cache")
-	    ssh  = "ssh #{configuration["user"]}@#{configuration["host"]}"
-	    cmd  = "find #{path} -type f -name \"*#{name}*\""
+      ssh  = "ssh #{configuration["user"]}@#{configuration["host"]}"
+      cmd  = "find #{path} -type f -name \"*#{name}*\""
 
-      system "#{ssh} '#{cmd}'"
+      to_run = "#{ssh} '#{cmd}'"
+
+      say to_run, :green if options[:verbose]
+      system to_run
 	  end
 
 	  desc "remove", "removes any matching cache files from the shared/cache directory"
@@ -28,8 +31,11 @@ module Karo
       ssh  = "ssh #{configuration["user"]}@#{configuration["host"]}"
       cmd  = "find #{path} -type f -name \"*#{name}*\" -delete"
 
+      to_run = "#{ssh} '#{cmd}'"
+
       if yes?("Are you sure?", :yellow)
-        system "#{ssh} '#{cmd}'"
+        say to_run, :green if options[:verbose]
+        system to_run
         say "Cache removed", :green
       else
         say "Cache not removed", :yellow