+++ lib/karo/db.rb
@@ -13,6 +13,7 @@ module Karo
     class_option :config_file, type: :string, default: Config.default_file_name,
                   aliases: "-c", desc: "name of the file containing server configuration"
     class_option :environment, aliases: "-e", desc: "server environment", default: "production"
+    class_option :verbose, type: :boolean, lazy_default: true, aliases: "-v", desc: "verbose"
 
     method_option :migrate, aliases: "-m", desc: "run migrations after sync", default: true
 	  desc "pull", "syncs MySQL database from server to localhost"
@@ -52,11 +53,19 @@ module Karo
       say "Loading #{options[:environment]} server database configuration", :green
 
       # Drop local database and re-create
-      system "mysql -v -h #{local_db_config["host"]} -u#{local_db_config["username"]} -p#{local_db_config["password"]} -e 'DROP DATABASE IF EXISTS `#{local_db_config["database"]}`; CREATE DATABASE IF NOT EXISTS `#{local_db_config["database"]}`;'"
+      to_run = "mysql -v -h #{local_db_config["host"]} -u#{local_db_config["username"]} -p#{local_db_config["password"]} -e 'DROP DATABASE IF EXISTS `#{local_db_config["database"]}`; CREATE DATABASE IF NOT EXISTS `#{local_db_config["database"]}`;'"
+
+      say to_run, :green if options[:verbose]
+
+      system to_run
 
       ssh  = "ssh #{configuration["user"]}@#{configuration["host"]}"
 
-      system "#{ssh} \"mysqldump --opt -C -u#{server_db_config["username"]} -p#{server_db_config["password"]} #{server_db_config["database"]}\" | mysql -v -h #{local_db_config["host"]} -C -u#{local_db_config["username"]} -p#{local_db_config["password"]} #{local_db_config["database"]}"
+      to_run = "#{ssh} \"mysqldump --opt -C -u#{server_db_config["username"]} -p#{server_db_config["password"]} #{server_db_config["database"]}\" | mysql -v -h #{local_db_config["host"]} -C -u#{local_db_config["username"]} -p#{local_db_config["password"]} #{local_db_config["database"]}"
+
+      say to_run, :green if options[:verbose]
+
+      system to_run
 
       if options[:migrate]
         say "Running rake db:migrations", :green
@@ -72,13 +81,23 @@ module Karo
 	  end
 
     desc "console", "open rails dbconsole for a given server environment"
+    method_option :tty, aliases: "-t", desc: "force pseudo-tty allocation",
+                  type: :boolean, default: true
     def console
       configuration = Config.load_configuration(options)
 
       path = File.join(configuration["path"], "current")
-      ssh  = "ssh #{configuration["user"]}@#{configuration["host"]} -t"
+      ssh  = "ssh #{configuration["user"]}@#{configuration["host"]}"
+
+      # Forces pseudo-tty allocation
+      ssh << " -t" if options[:tty]
+
       cmd  = "cd #{path} && bundle exec rails dbconsole #{options[:environment]} -p"
-      system "#{ssh} '#{cmd}'"
+
+      to_run = "#{ssh} '#{cmd}'"
+
+      say to_run, :green if options[:verbose]
+      system to_run
     end
 
 	end